feat(tictactoe): Update example to deploy
